requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
atari-py=0.2.6=pypi_0
blas=1.0=mkl
ca-certificates=2020.10.14=0
certifi=2020.11.8=py37haa95532_0
cffi=1.14.3=py37hcd4344a_2
click=7.1.2=pypi_0
cloudpickle=1.6.0=pypi_0
cudatoolkit=10.1.243=h74a9793_0
freetype=2.10.4=hd328e21_0
future=0.18.2=pypi_0
gym=0.17.3=pypi_0
intel-openmp=2020.2=254
joblib=0.17.0=pypi_0
jpeg=9b=hb83a4c4_2
libpng=1.6.37=h2a8f88b_0
libtiff=4.1.0=h56a325e_1
lz4-c=1.9.2=hf4a77e7_3
mkl=2020.2=256
mkl-service=2.3.0=py37h2bbff1b_0
mkl_fft=1.2.0=py37h45dec08_0
mkl_random=1.1.1=py37h47e9c7a_0
ninja=1.10.1=py37h7ef1ec2_0
nltk=3.5=pypi_0
numpy=1.19.2=py37hadc3359_0
numpy-base=1.19.2=py37ha3acd2a_0
olefile=0.46=py37_0
opencv-python=4.4.0.46=pypi_0
openssl=1.1.1h=he774522_0
pillow=8.0.1=py37h4fa10fc_0
pip=20.2.4=py37haa95532_0
protobuf=3.14.0=pypi_0
ptan=0.6=pypi_0
pycparser=2.20=py_2
pyglet=1.5.0=pypi_0
python=3.7.9=h60c2a47_0
pytorch=1.3.0=py3.7_cuda101_cudnn7_0
regex=2020.11.13=pypi_0
scipy=1.5.4=pypi_0
setuptools=50.3.1=py37haa95532_1
six=1.15.0=py37haa95532_0
sqlite=3.33.0=h2a8f88b_0
tensorboardx=2.1=pypi_0
tk=8.6.10=he774522_0
torchvision=0.4.1=py37_cu101
tqdm=4.54.0=pypi_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
wheel=0.35.1=pyhd3eb1b0_0
wincertstore=0.2=py37_0
xz=5.2.5=h62dcd97_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
